RF: merged tests for standard and framebyframe loading
